requirements.txt
taipy
requests
google-cloud-bigquery
python-dotenv